# GICv2 Distributor Config Cache